Master

+++ src/app/components/modals/new-project/new-project.component.ts
@@ -1,4 +1,12 @@
 import { Component, Inject, OnInit } from '@angular/core';
+import {
+  AbstractControl,
+  Form,
+  FormControl,
+  ValidationErrors,
+  ValidatorFn,
+  Validators,
+} from '@angular/forms';
 import { MatDialog, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { Router } from '@angular/router';
 import { Store } from '@ngrx/store';
@@ -16,9 +24,12 @@ import { ComposerUser } from 'src/models/user';
   styleUrls: ['./new-project.component.scss'],
 })
 export class NewProjectComponent implements OnInit {
+  projectNameControl: FormControl;
+  nftNameControl: FormControl;
   projectName: string = '';
   nftName: string = '';
   user: ComposerUser;
+  existingProjects: any[] = [];
 
   constructor(
     private store: Store<AppState>,
@@ -32,6 +43,12 @@ export class NewProjectComponent implements OnInit {
 
   ngOnInit(): void {
     this.user = this.userServices.getCurrentUser();
+    this.existingProjects = this.data.existingProjects;
+    this.projectNameControl = new FormControl(this.projectName, [
+      Validators.required,
+      this.forbiddenNameValidator(),
+    ]);
+    this.nftNameControl = new FormControl(this.nftName, [Validators.required]);
   }
 
   /**
@@ -39,36 +56,65 @@ export class NewProjectComponent implements OnInit {
    */
   public createProject() {
     if (this.projectName !== '' && this.nftName !== '') {
-      const project: NFTContent = {
-        ProjectId: Date.now().toString(),
-        ProjectName: this.projectName,
-        NFTName: this.nftName,
-        UserId: this.user.UserID,
-        TenentId: this.user.TenentId,
-        Timestamp: new Date().toISOString(),
-        CreatorName: '',
-        ContentOrderData: [],
-        NFTContent: {
-          BarCharts: [],
-          PieCharts: [],
-          BubbleCharts: [],
-          Tables: [],
-          Images: [],
-          Timeline: [],
-          ProofBot: [],
-          Stats: [],
-          CarbonFootprint: [],
-        },
-      };
-      sessionStorage.setItem('NFTCom', JSON.stringify(project));
-      this.store.dispatch(newProject({ nftContent: project }));
-      this.dndService.rewriteWidgetArr([]);
-      this.router.navigate([`/layout/home/${project.ProjectId}`]);
-      this.dialog.closeAll();
+      if (!this.checkIfAlreadyExists(this.projectName)) {
+        const project: NFTContent = {
+          ProjectId: Date.now().toString(),
+          ProjectName: this.projectName,
+          NFTName: this.nftName,
+          UserId: this.user.UserID,
+          TenentId: this.user.TenentId,
+          Timestamp: new Date().toISOString(),
+          CreatorName: '',
+          ContentOrderData: [],
+          NFTContent: {
+            BarCharts: [],
+            PieCharts: [],
+            BubbleCharts: [],
+            Tables: [],
+            Images: [],
+            Timeline: [],
+            ProofBot: [],
+            Stats: [],
+            CarbonFootprint: [],
+          },
+        };
+        sessionStorage.setItem('NFTCom', JSON.stringify(project));
+        this.store.dispatch(newProject({ nftContent: project }));
+        this.dndService.rewriteWidgetArr([]);
+        this.router.navigate([`/layout/home/${project.ProjectId}`]);
+        this.dialog.closeAll();
+      } else {
+        this.popupMsgService.openSnackBar(
+          'A project with the same name already exists '
+        );
+      }
     } else {
       this.popupMsgService.openSnackBar(
         'Please enter a project name and an NFT name'
       );
     }
+  }
+
+  /**
+   * @function checkIfAlreadyExists - check if the given project name already exists.
+   */
+  private checkIfAlreadyExists(newProjectName: string) {
+    let projectNameExists = false;
+    if (newProjectName !== '') {
+      for (let i = 0; i < this.existingProjects.length; i++) {
+        if (newProjectName === this.existingProjects[i].ProjectName) {
+          projectNameExists = true;
+          break;
+        }
+      }
+    }
+    return projectNameExists;
+  }
+
+  private forbiddenNameValidator(): ValidatorFn {
+    return (control: AbstractControl): ValidationErrors | null => {
+      const forbidden = this.checkIfAlreadyExists(control.value);
+      return forbidden ? { forbiddenName: { value: control.value } } : null;
+    };
   }
 }
